requirements.txt
pip
python-telegram-bot
logging
watson-developer-cloud